+++ delivery/src/test/java/com/nowayback/delivery/application/delivery/DeliveryServiceTest.java
@@ -1,11 +1,11 @@
-package com.nowayback.delivery.application;
+package com.nowayback.delivery.application.delivery;
 
-import com.nowayback.delivery.application.command.CreateDeliveryCommand;
-import com.nowayback.delivery.application.command.UpdateDeliveryRecipientInfoCommand;
-import com.nowayback.delivery.application.command.UpdateDeliveryStatusCommand;
-import com.nowayback.delivery.application.dto.DeliveryResult;
-import com.nowayback.delivery.application.exception.DeliveryApplicationErrorCode;
-import com.nowayback.delivery.application.exception.DeliveryApplicationException;
+import com.nowayback.delivery.application.delivery.command.CreateDeliveryCommand;
+import com.nowayback.delivery.application.delivery.command.UpdateDeliveryRecipientInfoCommand;
+import com.nowayback.delivery.application.delivery.command.UpdateDeliveryStatusCommand;
+import com.nowayback.delivery.application.delivery.dto.DeliveryResult;
+import com.nowayback.delivery.application.delivery.exception.DeliveryApplicationErrorCode;
+import com.nowayback.delivery.application.delivery.exception.DeliveryApplicationException;
 import com.nowayback.delivery.application.service.HubClient;
 import com.nowayback.delivery.domain.delivery.entity.Delivery;
 import com.nowayback.delivery.domain.delivery.repository.DeliveryRepository;
