+++ delivery/src/test/java/com/nowayback/delivery/presentation/DeliveryControllerTest.java
@@ -2,9 +2,8 @@ package com.nowayback.delivery.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nowayback.common.security.annotation.UserRole;
-import com.nowayback.delivery.application.DeliveryService;
-import com.nowayback.delivery.application.dto.DeliveryResult;
-import com.nowayback.delivery.application.exception.DeliveryApplicationException;
+import com.nowayback.delivery.application.delivery.DeliveryService;
+import com.nowayback.delivery.application.delivery.dto.DeliveryResult;
 import com.nowayback.delivery.domain.delivery.vo.DeliveryStatus;
 import com.nowayback.delivery.presentation.dto.request.CreateDeliveryRequest;
 import com.nowayback.delivery.presentation.dto.request.UpdateDeliveryRecipientInfoRequest;
